+++ Ch10Practice/package.bluej
@@ -2,42 +2,45 @@
 dependency1.from=PersonTester
 dependency1.to=Person
 dependency1.type=UsesDependency
+dependency2.from=Triangle
+dependency2.to=TriangleComponent
+dependency2.type=UsesDependency
 objectbench.height=100
 objectbench.width=658
 package.editor.height=400
-package.editor.width=560
-package.editor.x=589
-package.editor.y=472
-package.numDependencies=1
-package.numTargets=5
+package.editor.width=865
+package.editor.x=431
+package.editor.y=221
+package.numDependencies=2
+package.numTargets=6
 package.showExtends=true
 package.showUses=true
 project.charset=UTF-8
 target1.editor.height=700
 target1.editor.width=900
-target1.editor.x=271
-target1.editor.y=183
+target1.editor.x=792
+target1.editor.y=194
 target1.height=50
-target1.name=Coin
+target1.name=Triangle
 target1.naviview.expanded=false
 target1.showInterface=false
 target1.type=ClassTarget
 target1.typeParameters=
 target1.width=80
-target1.x=70
+target1.x=270
 target1.y=10
 target2.editor.height=700
 target2.editor.width=900
-target2.editor.x=287
-target2.editor.y=262
+target2.editor.x=471
+target2.editor.y=241
 target2.height=50
-target2.name=ButtonViewer
+target2.name=TriangleComponent
 target2.naviview.expanded=false
 target2.showInterface=false
 target2.type=ClassTarget
 target2.typeParameters=
-target2.width=100
-target2.x=160
+target2.width=140
+target2.x=370
 target2.y=10
 target3.editor.height=700
 target3.editor.width=900
@@ -54,27 +57,40 @@ target3.x=60
 target3.y=150
 target4.editor.height=700
 target4.editor.width=900
-target4.editor.x=893
-target4.editor.y=192
+target4.editor.x=271
+target4.editor.y=183
 target4.height=50
-target4.name=Triangle
+target4.name=Coin
 target4.naviview.expanded=false
 target4.showInterface=false
 target4.type=ClassTarget
 target4.typeParameters=
 target4.width=80
-target4.x=270
+target4.x=70
 target4.y=10
 target5.editor.height=700
 target5.editor.width=900
-target5.editor.x=490
-target5.editor.y=307
+target5.editor.x=287
+target5.editor.y=262
 target5.height=50
-target5.name=Person
+target5.name=ButtonViewer
 target5.naviview.expanded=false
 target5.showInterface=false
 target5.type=ClassTarget
 target5.typeParameters=
-target5.width=80
-target5.x=70
-target5.y=70
+target5.width=100
+target5.x=160
+target5.y=10
+target6.editor.height=700
+target6.editor.width=900
+target6.editor.x=490
+target6.editor.y=307
+target6.height=50
+target6.name=Person
+target6.naviview.expanded=false
+target6.showInterface=false
+target6.type=ClassTarget
+target6.typeParameters=
+target6.width=80
+target6.x=70
+target6.y=70
